// src/schema/types/fields.ts
export interface UIElementBase {
    order?: number;
}

export interface DbOptions {
    index?: boolean;
    unique?: boolean;
    nullable?: boolean;
}

export interface Base_OPT<T = unknown> extends UIElementBase {
    _meta: 'field';
    name: string;
    label: string;
    placeholder?: string;
    info?: string;
    required?: boolean;
    default?: T;
    disabled?: boolean;
    validate?: (value: any) => string | void;
    db?: DbOptions;
    persist?: boolean;
}
// ...
export interface Ref_OPT extends Base_OPT<string | number> {
    type: 'ref';
    ref: string;
    alias?: string; // <-- Новое свойство для явного контроля имени
    relation?: {
        type: 'many-to-one' | 'one-to-one';
    };
}